+++ Classes/Nakama/Public/NakamaSDK/NMatchPresence.h
@@ -30,10 +30,10 @@ namespace Nakama {
 		NMatchPresence(MatchPresence message)
 		{
 			id = message.match_id();
-			for (size_t i = 0; i < message.joins_size(); i++) {
+			for (int i = 0; i < message.joins_size(); i++) {
 				join.push_back(NUserPresence(message.joins(i)));
 			}
-			for (size_t i = 0; i < message.leaves_size(); i++) {
+			for (int i = 0; i < message.leaves_size(); i++) {
 				leave.push_back(NUserPresence(message.leaves(i)));
 			}
 		}
